Updated ReadMe.md to include some tips/hints (#34)
* Updated ReadMe.md to include some tips/hints

As I've talked with @Tyler, it should be clear that since this requires hardware interaction, the module must be installed with high privileges. 
I've also indicated the path to the module as an "hint". I assumed that it will always be the same since it's installed as root (correct me if I'm wrong).

* Update README.md

As @Tyler suggested, removed the previously added hint.
Added the `sudo pwsh` before the Install-Module